30-agent/requirements.txt

PyPDF2==3.0.1
python-docx==1.1.0
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.31.0
